feat: Add testing infrastructure
Local testing and development infrastctureinfrastructure is now
generated with terraform. This opens up the possibility of locally
iterating on cluster manifests before pushing to the main cluster or
running automated tests on manifest changes. Integration into github
CI workflows is also planned.

The reason that the testing_registry resource is provisioned by
ansible, rather than the simpler option of provisioning a docker
container for each registry server is to replicate the behaviour that
would be on the production server (Running as a LXC container on
proxmox).