feat: improve talos patch generator
The patch generator now outputs it's generated jsonpatches to the .gen
directory. It also allows the user to specify a kubernetes
version (required due to a bug with talos and kubernetes 1.24) and
automatically merges the generated talosconfig with the user's one.

A bug with how talosconfig endpoints were passed was fixed, and the
talos machineconfig endpoint was changed to an IP to workaround DNS
rate limiting on my home network (pihole defaults).