feat: patch generator supports multiple environments
In order to effectively test the terraform configuration I need a way
to run it with testing values. The original ansible playbook used to
have a hardcoded limitation on a single .gen path. Now the `project`
variable controls which environment (test|prod|dev) to deploy.

Registry endpoints are also no longer hardcoded.