feat: Add storage provider creation to kubernetes cluster
The cluster will now, once created, be initialised with a deployment
of OpenEBS's Mayastor. It is pinned to a certain version, and runs
with a Helm chart deployment of etcd. Etcd's storage is provisioned on
every worker node and consists of directories at
/var/local/mayastor. The path here should be related to the one in the
patch generation playbook. Other improvements would be creating
storage pools automatically and removing the awful hack in the
"kubectl_manifest" "mayastor_yamls" resource.