feat: add support for multiple environments
In order to effectively test the terraform project I need a way of
testing how it deploys to physical proxmox machine, without
interrupting service of the main cluster. To alleviate this a
serperate set of variables can now be defined to ensure hostnames, ip
addresses, disks and vmids don't conflict.

This commit also includes a fix that stops the controlplane and worker
base yamls from being destroyed whenever the patches need to be
regenerated.