feat: Add NFS server to cluster_containers module
The NFS server container is provisioned in the same way as the
registry container, and has the same issues regarding local-exec
provisioners. Unfortunately there isn't an easy way to cloud-init the
ubuntu cloud images for proxmox LXC containers. Hopefully it is
something the proxmox team can add in the future. In the future I will
probably build my own images and use them.