feat: talos manifest generation playbook uses values from terraform
The playbook used to have a seperate set of variables that were not
derived from the terraform state/configuration. This caused issues
whenever changing values in each aspect of the project. It also meant
I needed to commit a lot of hardcoded values. This commit templates
the varfiles in terraform and passes those vars to ansible to continue
manifest/patch generation as usual.

TODO: Document newly added variables in the talos_proxmox_cluster and
root modules. And figure out a repetable way of deploying the Mayastor
manifests without commenting and uncommenting code (targets are
probably the way to go).