feat: initially add terraform provider for creating talos cluster.
The provisioner based approach in the main terraform project is hacky
and does not properly manage the lifecycle of talos. It also requires
an additional binary dependency. The idea of this provider would be
for it to directly interact with the talos API, and generate the
required talosconfigs from templates. It's also a learning exercise
for go, a new language for me.